Add documentation and fix implicit conversion.

+++ include/vtzero/tile.hpp
@@ -37,6 +37,11 @@ namespace vtzero {
 
     } // namespace detail
 
+    /**
+     * This class repesents a tile, it doesn't contain any data, just the
+     * zoom level, x and y coordinates and the extent. The class is used to
+     * convert tile coordinates into web mercator coordinates.
+     */
     class tile {
 
         uint32_t m_x;
@@ -46,6 +51,14 @@ namespace vtzero {
 
     public:
 
+        /**
+         * Construct a tile.
+         *
+         * @param x X coordinate of the tile (0 <= x < 2^zoom)
+         * @param y Y coordinate of the tile (0 <= y < 2^zoom)
+         * @param zoom Zoom level of the tile (0 <= zoom < 16)
+         * @param extent Extent used in vector tiles
+         */
         tile(uint32_t x, uint32_t y, uint32_t zoom, uint32_t extent = 4096) noexcept :
             m_x(x),
             m_y(y),
@@ -56,29 +69,49 @@ namespace vtzero {
             vtzero_assert_in_noexcept_function(y < detail::num_tiles_in_zoom(zoom) && "y coordinate out of range");
         }
 
+        /// The x coordinate.
         uint32_t x() const noexcept {
             return m_x;
         }
 
+        /// The y coordinate.
         uint32_t y() const noexcept {
             return m_y;
         }
 
+        /// The zoom level.
         uint32_t zoom() const noexcept {
             return m_zoom;
         }
 
+        /// The extent.
         uint32_t extent() const noexcept {
             return m_extent;
         }
 
+        /**
+         * Transform the vector tile coordinates specified as argument into
+         * web mercator coordinates
+         *
+         * This uses integer arithmetic only.
+         *
+         * @param p Vector tile coordinates
+         * @returns coordinate pair of web mercator coordinates (in millimeters).
+         */
         std::pair<int64_t, int64_t> transform_int(const point p) const noexcept {
             const int64_t d = static_cast<int64_t>(detail::num_tiles_in_zoom(m_zoom)) * m_extent;
-            const int64_t x = 2 * detail::max_coordinate_epsg3857_mm * (m_extent * m_x  + p.x) / d - detail::max_coordinate_epsg3857_mm;
-            const int64_t y = 2 * detail::max_coordinate_epsg3857_mm * (m_extent * m_y  + p.y) / d - detail::max_coordinate_epsg3857_mm;
+            const int64_t x = 2 * detail::max_coordinate_epsg3857_mm * (static_cast<int64_t>(m_extent * m_x) + p.x) / d - detail::max_coordinate_epsg3857_mm;
+            const int64_t y = 2 * detail::max_coordinate_epsg3857_mm * (static_cast<int64_t>(m_extent * m_y) + p.y) / d - detail::max_coordinate_epsg3857_mm;
             return {x, y};
         }
 
+        /**
+         * Transform the vector tile coordinates specified as argument into
+         * web mercator coordinates
+         *
+         * @param p Vector tile coordinates
+         * @returns coordinate pair of web mercator coordinates (in meters).
+         */
         std::pair<double, double> transform_double(const point p) const noexcept {
             const auto r = transform_int(p);
             return {static_cast<double>(r.first) / 1000,
